fix: remove unnecessary u64 cast in billing.rs
clippy lint: casting to the same type (u64 -> u64) with BILLING_PERIOD_SECONDS
which is already typed u64. This fixes the Lightning Adapter Verification CI.

Closes #222

+++ internal/engine/src/billing.rs
@@ -524,7 +524,7 @@ mod protocol_fee_tests {
     fn test_period() -> BillingPeriod {
         BillingPeriod {
             start_unix: 0u64,
-            end_unix: BILLING_PERIOD_SECONDS as u64,
+            end_unix: BILLING_PERIOD_SECONDS,
             deployment: GatewayDeployment::Managed,
         }
     }
